Damage is Shown
Damage is sort of shown (thought it goes outward in instead of right to left.

My code my be difficult to understand since I don't make a ton of methods, but feel free to edit as you please. Might be better to use physics2d to check collisions of the Unity API methods to make sure objects 'fight' when next to each other.