Things can attack
Enemies and allies battle (kind of). Animations are a little weird.

Health moves as damage is taken.

Repositioned UI to make it look a bit neater.

Animation control is a off, but we can figure that out later. When a target dies, ally or enemy will move to next closest target.